refactor(pickers): split remaining picker surfaces (#529)
Problem: the checks, CI, and release picker flows still lived inside
pickers.lua beside the public wrappers and convenience helpers. That left
multiple unrelated surfaces coupled to one file even after the PR and
issue picker split.

Solution: move the checks, CI, and release pickers into dedicated
modules and reduce pickers.lua to thin wrapper functions and the small
public helpers that remain user-facing.
