+++ spec/pickers_spec.lua
@@ -21,8 +21,11 @@ local preload_modules = {
   'forge.logger',
   'forge.ops',
   'forge.picker',
+  'forge.picker.checks',
+  'forge.picker.ci',
   'forge.picker.issue',
   'forge.picker.pr',
+  'forge.picker.release',
   'forge.picker.shared',
   'forge.surface_policy',
   'forge.state',
@@ -34,8 +37,11 @@ local loaded_modules = {
   'forge.logger',
   'forge.ops',
   'forge.picker',
+  'forge.picker.checks',
+  'forge.picker.ci',
   'forge.picker.issue',
   'forge.picker.pr',
+  'forge.picker.release',
   'forge.picker.shared',
   'forge.picker.session',
   'forge.pickers',
